requirements.txt
# 核心库
python-dotenv
pandas
notebook

# 金融数据源
yfinance
tushare             # <-- 已替换

# GUI框架 (后续部分使用)
gradio

# 可视化 (后续部分使用)
plotly

# LLM SDKs (后续部分使用)
google-generativeai
openai
dashscope